requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
databricks-sdk==0.36.0
openai==1.51.0
psycopg2-binary==2.9.9
fpdf2==2.8.1
pydub==0.25.1
httpx==0.27.2
